requirements.txt
aiogram
aiohttp
asyncpg
pandas
python-dotenv
cachetools